Bump up PyYaml to recommended version to remediate vulnerability - DROP PYTHON 3.4 SUPPORT (#366)
Summary
-------

-  Removing PyYaml from SDK test.txt
-  Dropping  Python 3.4

Dependabot found critical vulnerability in this dependency and it is not needed in the SDK
Python 3.4 no longer supports required libraries to support new and existing features

Note: PyYaml was previously an explicit dependency, however PyYaml is also a transient dependency in python-coverals, which has already upgraded to the latest version of PyYaml. This version of PyYaml no longer supports Python version 3.4. For this reason we have chose to also drop Python 3.4, as we need to continue proper code coverage on our SDKs to ensure the highest quality of code.

Test plan
---------
- FSC

Issues
------

-  “OASIS-8054"